backend/requirements.txt
# FastAPI и зависимости
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.5.2
email-validator==2.2.0

# База данных
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1


# Google Gemini AI (новый SDK)
google-genai>=1.0.0

# RAG и векторные операции
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==1.26.4

# Утилиты
python-dotenv==1.0.1
